#210 - moved json response from another module
